Move port types into factory to reduce number of new classes.
